Commit 2
in the new version, I used an array list to store to store the vehicles and make it easy to carry out the iterations. I added a method to compute details and rental costs for each vehicle. I added additional computations to show the cost of renting a type of vehicle on various days, and a display text to thank the user for their service.